refactor embeddings
